requirements.txt
Flask
requests
pandas
Pillow
